% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-assessment.R
\name{dist_cie2000}
\alias{dist_cie2000}
\title{CIE 2000 Distance between a set of colors}
\usage{
dist_cie2000(x)
}
\arguments{
\item{x}{vector (or matrix) of colors.}
}
\value{
distance matrix of the colors
}
\description{
The CIE 2000 distance between to two colors is, similar to the CIE 76 distance,
based on the LAB color space, but additional adjustments are made to better
adjust for non-linearities in the human perception of color (we see different hues of color at
different resolutions).
}
